requirements.txt
asn1crypto==0.24.0
asyncio==3.4.3
certifi==2019.3.9
cffi==1.12.2
cryptography==2.6.1
fakeRPiGPIO==0.3a1
future==0.17.1
idna==2.8
Markdown==3.0.1
pexpect==4.6.0
pip==19.0.3
ptyprocess==0.6.0
pycparser==2.19
pyghmi==1.2.16
setuptools==40.8.0
six==1.12.0
telnetlib3==1.0.2
virtualenv==16.4.3
wheel==0.33.1
wincertstore==0.2
